refactor: extract ConfirmationResolver and TurnDispatch from MessageRouter
